+++ frontend/src/routes/login/route.lazy.tsx
@@ -1,7 +1,17 @@
-import { createLazyFileRoute } from "@tanstack/react-router";
-
-import { Login } from "./route";
+import { Button } from "@mantine/core";
+import { createLazyFileRoute, Link } from "@tanstack/react-router";
 
 export const Route = createLazyFileRoute("/login")({
   component: Login,
 });
+
+export function Login() {
+  return (
+    <>
+      <p>Login Page</p>
+      <Link to="/">
+        <Button>Top</Button>
+      </Link>
+    </>
+  );
+}
