+++ frontend/src/routes/login/route.tsx
@@ -1,10 +1,8 @@
-import { Button } from "@mantine/core";
-import { Link, createFileRoute } from "@tanstack/react-router";
+import { createFileRoute } from "@tanstack/react-router";
 
 import { BACKEND_EJS_VARIABLE_KEYS } from "@/staticDataRouteOption";
 
 export const Route = createFileRoute("/login")({
-  component: Login,
   staticData: {
     openGraph: {
       title: `<%= ${BACKEND_EJS_VARIABLE_KEYS.siteName} + ' | Login' %>`,
@@ -13,14 +11,3 @@ export const Route = createFileRoute("/login")({
     },
   },
 });
-
-export function Login() {
-  return (
-    <>
-      <p>Login Page</p>
-      <Link to="/">
-        <Button>Top</Button>
-      </Link>
-    </>
-  );
-}
