+++ frontend/src/routes/route.tsx
@@ -1,12 +1,8 @@
-import { Button, Group } from "@mantine/core";
-import { Link, createFileRoute } from "@tanstack/react-router";
-
-import { DeleteParrot } from "./-components/DeleteParrot";
+import { createFileRoute } from "@tanstack/react-router";
 
 import { BACKEND_EJS_VARIABLE_KEYS } from "@/staticDataRouteOption";
 
 export const Route = createFileRoute("/")({
-  component: Home,
   staticData: {
     openGraph: {
       title: `<%= ${BACKEND_EJS_VARIABLE_KEYS.siteName} %>`,
@@ -15,16 +11,3 @@ export const Route = createFileRoute("/")({
     },
   },
 });
-
-function Home() {
-  return (
-    <>
-      <DeleteParrot />
-      <Group justify="center" mt="lg">
-        <Button component={Link} to="/login">
-          LOGIN
-        </Button>
-      </Group>
-    </>
-  );
-}
